KVM: x86: Track kvm_vcpu_arch.pending_ioapic_eoi as a u8 to drop dead BUG_ON()
Track the vector associated with KVM_EXIT_IOAPIC_EOI as a u8, as the vector
is zero-initialized and is only ever set to a legal vector.  I.e. unlike
many of the other vector variables in KVM, it can't be -1 (and can't be
greater than 255, though that should hold true for all vectors in KVM).

Drop the now fully dead BUG_ON(), which aside from being completely
unnecessary, was also useless, as it wouldn't guard against a negative
value, i.e. wouldn't detect KVM's magic -1 value.

For all intents and purposes, no functional change intended.

Suggested-by: Paolo Bonzini <pbonzini@redhat.com>
Link: https://patch.msgid.link/20260806171400.4119301-1-seanjc@google.com

+++ arch/x86/include/asm/kvm_host.h
@@ -974,7 +974,7 @@ struct kvm_vcpu_arch {
 		bool pv_unhalted;
 	} pv;
 
-	int pending_ioapic_eoi;
+	u8 pending_ioapic_eoi;
 	int pending_external_vector;
 	int highest_stale_pending_ioapic_eoi;
